Fix `ApplyFiltersButton` offset (#213)
This PR fixes a bug where clicking the `ApplyFiltersButton` did not reset the pagination offset. Also, it includes a fix that I made in the `react-16.8` branch to only include recent searches screenreader counts when on a universal page.

J=SLAP-2202
TEST=manual

Checked that previously, navigating to the second page on the `people` vertical and selecting a filter with the apply filters button would not reset the offset, but after the fix, it resets to the first page.