Fix SearchBar stories (#214)
This PR fixes the `SearchBar` stories so they display recent searches as expected. Adding recent searches manually to the browser's local storage was causing problems. This is likely because of the hard-coded timestamp, which may cause the data to be removed from storage if it is too far in the past. I changed the recent searches back to being added using `userEvent`s.

J=none
TEST=auto

See that Percy snapshots include the recent searches for `SearchBar` stories that have the dropdown expanded.