Display counts for StandardFacets (#226)
Update `StandardFacets` to display the number of results associated with each facet option. The results count is not displayed for `NumericalFacets` or `HierarchicalFacets`. Update Jest tests to match the new facet option labels.

J=SLAP-2224
TEST=auto, manual

Check that the updated Jest tests pass and see in the test-site and Storybook that `StandardFacets` displays the count next to each option.