Show pagination info in results count (#228)
This PR adds a range to the `ResultCount` that describes the range of results currently displayed by `Pagination`.  For example `11-30 of 45 Results`.  Since this range describes ranges from pagination, it only applies to `VerticalSearch` pages and does not show on `UniversalSearch` pages.

Old jest tests for behavior on no results were edited.

J=SLAP-2225
TEST=manual, auto

Vertical Searches were performed on the test site and pagination was used to assess if proper ranges were displayed in the Result Count message.
Jest tests were written to test that pagination range is displayed when pagination occurs, and is not displayed when pagination is not necessary.