fix ThumbsFeedback issues (#233)
This PR fixes ThumbsFeedback issue reported by product of not having a default width and height, as well as rotation issue with the thumbs-down button. These issues are fixes by making cssClasses (now renamed to customCssClasses for consistency) optional and making builtinCssClasses default. ThumbsFeedback builtinCssClasses were originally set as default in and was used as part of DirectAnswer and StandardCard component, which caused the issues in product's use case as they were using it as a standalone component. Now the builtinCssClasses will render when ThumbsFeedback is used as a standalone component.

J=SLAP-2228
TEST=manual

Tested by simulating product's use case on the test site and making ThumbsFeedback a standalone component. The styling is rendered as expected.