fix inputPrefix collision issue (#232)
This PR fixes inputPrefix collision issue in RangeInput. The padding with inputPrefix is increased to avoid potential collision of max(min) input with the inputPrefix for up to three characters. The max/min input is also aligned to the right as approved by product.

J=SLAP-2212
TEST=manual

Tested the change by adding an inputPrefix of three characters (USD) to NumericalFacets. With padding increased, the max/min input no longer collides with inputPrefix. The min/max input was aligned to the right because of aesthetic reason approved by product - so without any input, the min/max placeholder and inputPrefix would not appear cramped together - input numbers appeared as expected with the right alignment.