fix tooltip location (#234)
with rangeInput having `w-full`, dynamic width sizing, the manual top/right position styling adjustment no longer work if user change the width of rangeInput. The tooltip would overlap with the input boxes. Updated to ensure the tooltip always snap to the end of the rangeInput box
![Screen Shot 2022-07-13 at 4 45 03 PM](https://user-images.githubusercontent.com/36055303/178831055-c8ce3d6e-abe8-4eda-82dc-01d736ae1921.png)

J=none
TEST=manual&auto

see that tooltip is placed at the end of the rangeInput box no matter the sizing
w-56 sizing
![Screen Shot 2022-07-13 at 4 47 54 PM](https://user-images.githubusercontent.com/36055303/178831551-d41ad504-f9e5-49d0-8963-f9a398bff71d.png)
span the full page
![Screen Shot 2022-07-13 at 4 48 11 PM](https://user-images.githubusercontent.com/36055303/178831554-bbb3a232-f623-467b-8893-d653e791e5bc.png)