Reset facets on searches from FilterSearch (#236)
This PR fixes a bug where facets were not being reset when applying a filter from `FilterSearch` with `searchOnSelect=true`. Product wants the behavior to be like conducting a search from the search bar, where facets are cleared. I also moved the `setOffset` call so it is only made when a search is executed, not whenever a filter is selected (e.g. when `searchOnSelect=false`).

J=SLAP-2240
TEST=auto, manual

Check that the updated Jest tests pass. In the test-site, see that before the changes, facets remained when selecting a filter, even if that sometimes led to no results being shown. After the changes, see that facets are removed when selecting a filter from `FilterSearch`.