Update unset-value transitive dependency.
This transitive dependency of Storybook had a vulnerability identified
by Snyk. I manually updated the version we use in the package-lock. This
new version of unset-value does not have the vulnerability. Even though
I updated unset-value by a major, I looked at the release notes and we
will not be impacted by the breaking change.

TEST=manual

Built and served the Storybook site. Everything worked as expected.