make option counts optional (#248)
This PR provides users the option to turn off option counts on standard filters.  An optional prop, `showOptionCounts` was added to the `StandardFacets` prop interface.  If unspecified, it defaults to true.

J=SLAP-2251
TEST=auto, manual

It was tested that when `showOptionCounts` was set to false, option counts were not displayed. When it was set to true, or omitted entirely, option counts were displayed.
A jest test was written to show that when `setOptionsFilters` is false, then the counts do not display. 
A Storybook story was added that displayed standard facets without option counts.