Make CardComponent and CardProps generics (#231)
Update `CardProps` to take an optional generic type param to pass to `Result` and update `CardComponent` to accept one to pass to `CardProps`. The default value is the same as that of the generic type param for `Result`.

J=SLAP-2232
TEST=manual

Check that the type param is not required (e.g. `StandardCard` works like normal). Start creating a custom card that specifies a type for `CardProps` and see that we get compile-time type checking for the `rawData` of the result in the component body.