remove initialValue in Dropdown (#249)
InitialValue is not use by FilterSearch or SearchBar. In the case of user want to set an initial value for the search bar, they would set it through answersActions.setQuery() function, which updates on second render, whereas initialValue is set in useState call in useInputContextInstance on first render. This behavior is taken care of by `parentQuery` field already.

J=SLAP-2249