Add showMoreLimit prop (#245)
Add a `showMoreLimit` prop to `StandardFacet`. When the number of options exceeds this limit, display only that many options, a searchable facets input, and a show more button. The show more button, when clicked displays all the options. If the number of options is less than the limit, show all the options and no search bar

J=SLAP-2252
TEST=auto

Added two new stories: one to show the prop at work, and the second to show what happens when the show more button is pressed