fix the issue of filter search dropdown pushing content down (#258)
regression from switching from absolute to relative in [previous pr](https://github.com/yext/search-ui-react/pull/205) when changing filter search styling so the width of the dropdown menu match the width of the input box. However this cause the filter dropdown menu to push the content down when it's active. Setting z-index with relative on an element inside that block will only alter its layer with respect to other elements inside the same block.

This pr update the styling back to using absolute but set the dropdown menu to be relative to the filter search container and use `w-full` to span its width such that it match the width of the input box.

J=SLAP-2282
TEST=manual
Before and after:
<img width="256" alt="Screen Shot 2022-07-26 at 10 41 27 AM" src="https://user-images.githubusercontent.com/36055303/181036226-1f992d9a-8c3a-4aa8-b3c0-0290e835649b.png"><img width="272" alt="Screen Shot 2022-07-26 at 10 40 16 AM" src="https://user-images.githubusercontent.com/36055303/181035936-b8e54908-53d5-4891-9acf-04b5e78589a8.png">