unique keys for applied filter pills
I assumed that displayNames would be unique for NLP filters coming from
the API. This is not necessarily the case, and results in duplicate keys
which can result in UI bugs.

Including an underscore and index at the end makes the keys effectively unique barring
very weird edge cases.

J=none
TEST=manual

searching for Richmond on a locations page (does not exist yet, will be introduced in Alex's maps pr)
then searching for texas, does not result in an incorrect UI state