Added Dropdown Stories (#288)
Added stories for `Dropdown` component. Though it is used in `FilterSearch` and `SearchBar`, we figured it'd be good to have separate stories because users should be able to use `Dropdown` for their own components. `Dropdown` doesn't have any built-in css, so I used the `FilterSearch` css as a reference for styling.

On writing the tests, I found a wcag bug with `DropdownInput` that wasn't setting the `aria-activedescendant` correctly on select. I had to set `focusedIndex` back to `-1` after performing actions related to pressing 'enter'. 

J=SLAP-2320
TEST=auto