fix node v18 visual test issue (#293)
This PR fixes the visual coverage test in node v18, as Run Tests / call_run_tests-react-16 / tests (18.x) from github actions in the past pull requests have been failing. It is a temporary fix to allow the visual test in node 18 to run. It is not a recommended approach to use node option with OpenSSL 3 in the legacy mode, as it would be running with known insecure algorithms. But for testing purpose, there would not be security issue. This PR only sets node option if the node version is 17+. Since issue is caused by storybook using a lower version of its dependency webpack (with newest version supporting node 18), it will be fixed if storybook bump their webpack version.

TEST=auto