Static filter state updates (#295)
Update the components that interact with the static filters State to work with the changes in Headless. The `StaticFilters` and `AppliedFilters` components will only work with field value filters. If someone wants to create and display more complex static filters, they should fork the components and do that themselves.

Also, because the Headless-React alpha versions include the change in Core to make direct answers generic, I updated the component to manually cast those values as strings for now. That behavior should be changed with the other direct answer work.

J=SLAP-2327
TEST=manual

Spin up the test-site and see that static filters, range input filters, and filter search filters can be successfully applied and removed.