FieldValueDirectAnswer component (#296)
Added FieldValueDirectAnswer to handle all the built in field types for field value direct answer. Note, we currently do not support internationalization so addresses will be written as if they are US addresses.

in nidhi's PR for featured snippet direct answer, she will implement the logic to handle rich text field type using react-markdown library. So, the switch case for rich_text in FieldValueDirectAnswer will be updated once that is done. Along with the tests for rich_text.

J=SLAP-2332
TEST=auto

added storybook and jest tests for the different field types in field value direct answer