Add FeaturedSnippetDirectAnswer component (#297)
This PR adds a `FeaturedSnippetDirectAnswer` component that is used inside of `DirectAnswer` to render featured snippet direct answer types. Another item will add support and tests for rich text conversion (SLAP-2340), so this PR handles the `MultiLineText` field type.

J=SLAP-2335
TEST=auto, manual

See that the added Jest tests pass and the Percy snapshot for "plain text featured snippet" is generated that matches the previous "featured snippet" snapshot. Spin up the test-site and see that "where was joe exotic born" returns a multi-line text featured snippet and displays it as expected.