handle "unknown" field type in field value DA (#299)
Update DirectAnswer component to accept an optional prop field `UnknownFieldTypeDisplay` to handle rendering results of "unknown" field type in field value direct answer. If not provided, FieldValueDirectAnswer will use the default component, which would attempt to render result.value as a string with toString() method. It would console log a warning for user if it encounter results of "unknown" field type where value is not of type string or number. 

J=SLAP-2337
TEST=auto&manual

Added jest tests for handling unknown field type in field value through custom and default UnknownFieldTypeDisplay component.
Also manually tested with a UnknownFieldTypeDisplay component in Universal page and with a query "what is yen banana price" for a fieldApiName "c_bananaPrice"