Further updates to account for changes in `SearchHeadlessProvider`. (#302)
* Further updates to account for changes in `SearchHeadlessProvider`.

The README was updated to reflect the fact that `SearchHeadlessProvider` now
takes in a Headless instance. The README also shows someone how to handle
acquisition of a session tracking ID. I added session tracking logic to the
test-site as well as a further reference.

TEST=auto

Ensured that all unit and WCAG tests passed locally.