update readme "Getting Started" section (#304)
update readme wording now that `SearchHeadlessProvider` requires a `SearchHeadless` instance instead of directly passing in the credentials. Also include small changes such that the code sample works out of the box.

TEST=manual

copy and pasted the code sample to test-site App.tsx using the actual credentials, See that it works as expected.