Only `e.preventDefault` when Dropdown is active (#307)
Update the `Dropdown` component to only call `e.preventDefault` for the up/down arrow keys if the dropdown is active.

J=SLAP-2351
TEST=manual

Spin up the test-site, load results, and see that the up and down arrow keys scroll the page when the dropdown is closed. When the dropdown is active, up/down arrows work as expected to navigate through dropdown options. Tested both the `SearchBar` and `FilterSearch` dropdowns.