use createRoot when possible for rendering pin components
This PR uses a dynamic imports with a .catch fallback for when
react-dom/client does not exist.

MapboxMap code had to be tweaked due to differences between ReactDOM.render
and ReactDOM.createRoot().render(). Namely, createRoot() does not modify the
container element unless it already exists on the physical page.
This means that we can't call document.creatElement('div'), use createRoot,
and THEN attach it to the mapbox map. Instead, we have to attach the div to the
map first.

This also seems it make it more difficult to unit test, my guess would be because
createRoot does not do anything on the server. For now I use some jest.mocks to
ensure the right methods are being called in the right environments.

TEST=manual,auto