Export CSS bundle without Tailwind preflight (#322)
Add an export for a CSS bundle without the Tailwind preflight CSS resets. While testing, I noticed that some of our components rely on Tailwind's default border color, which is set in the CSS resets. For users who use the new bundle and specify the Tailwind resets as custom resets, the default border color isn't found because they aren't using Tailwind. I updated the components with borders that were missing a border-color so the styling is consistent regardless of what CSS resets (if any) a user provides. I also updated the default border color in the Storybook and test-site Tailwind configs to match what a no-resets user would see if we forgot to include a border-color for a border in the future.

J=SLAP-2418
TEST=manual

Remove Tailwind from the test-site and instead import the new bundle without Tailwind preflight. Add the Tailwind preflight CSS resets to `test-site/src/index.css` and see that the components' styling matches how they look with Tailwind (or with the CSS bundle with preflight). Scope the Tailwind CSS resets with a class name and see that only parts of the app with that class name have the CSS resets applied, while the rest of the app is unaffected.