FilterSearch updates for `builtin.location` (#343)
Update `FilterSearch` to account for special backend logic when searching on the `builtin.location` field. In this case, the backend can return filters on `builtin.region` or `address.countryCode` as well, so these `fieldId`s are also considered "matching" for the component. Filters on these fields can be used to populate the input and should be unselected when a new filter is selected from `FilterSearch`.

J=SLAP-2495
TEST=auto, manual

See that the added Jest tests pass. Spin up the test-site and add static filters on `builtin.region`. See that they are used to populate the `FilterSearch` input, warnings are logged, and the filters are unselected by `FilterSearch` as expected.