Update Geolocation to remove all location fields (#344)
Update `useGeolocationHandler` so that it replaces all location filters with the new filter. In other words, filters on the `builtin.location`, `builtin.region`, and `address.countryCode` fields are all removed when a new filter is selected via the `Geolocation` component. This matches the behavior of `FilterSearch` and ensures the `State` doesn't end up with multiple location filters applied at the same time, which can have unpredictable results from the backend.

J=SLAP-2495
TEST=auto

See that the Jest tests still pass even when the mocked state includes other kinds of location filters besides those on `builtin.location`.