add checks to fail combined-coverage tests after each part (#346)
Currently tests are failing silently in Github actions. This PR adds checks to combined-coverage.sh script to exit if a test fails. The three tests currently failing on main will now be caught, I created this item to handle them separately, as it will require a change to make react versions to test programmable in reusable workflows

J=SLAP-2459
TEST=manual

ran actions with an intentional failure (log). Also commented out npm publish from the slap publish command and tested that it makes it passed the check there