PagesJS compatibility (#400)
Introduce tsup and improve the package.json file in order to support PagesJS.

This works in both dev and prod mode of pagesJS. Currently the storybook tests are failing, however we will address that in a future PR.

This changes updates our esm files to created in a bundle in the dist folder. The esm files now end with .mjs rather than .js which may be a breaking change for some build tools. I struggled to get `type: module` to work in many situations, whereas using .mjs files worked better across the board. Regardless, these changes caused a lot of different issues, so I think this should go in the next major version

J=BACK-2545
TEST=manual

Publish an alpha version and test it with @yext/pages@1.0.0-rc.3. Add several search components to the page and confirm they work as expected.

Currently the storybook tests are failing and need to be fixed.