Function Vertical Support via search-core/headless upgrade (#459)
This change upgrades the version for search-core and search-headless to catch the changes which add support for function verticals. It also adds a function vertical page to the test-site for easier testing. An image of that page is attached.

J=WAT-3474
TEST=auto,manual

Tested via test-site, saw responses returned were upper cased as expected by the plugin, and the source was FUNCTION_VERTICAL.

<img width="1603" alt="Screenshot 2024-08-29 at 4 03 26 PM" src="https://github.com/user-attachments/assets/14e939eb-0ac3-493f-be83-44fdf9350358">