ksearch: adjust id generation to add basename from callers (#471)
This PR adjusts our useId function, which generates ids for
html elements, to also take a 'basename' argument from
callers. This is to help reduce the likelihood of an id
made conflicting with an id somewhere else on the page.

J=WAT-4464
TEST=auto

Ran tests. Was having issues getting the test site running

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>