Allow disabling built-in CSS classes for FilterSearch component
To address https://github.com/yext/search-ui-react/issues/481
Adds a new flag to FilterSearchProps named disableBuiltInClasses.
If this property is true, and there are customCssClasses set, then
we will completely remove the builtin CSS classes from the
component.

TEST=manual
Spun up test site and set the property on a component. Saw the
classes get replaced as expected.