search-ui-react: Resolve build errors
The `npm run build` command was failing due to errors with the tsup
build command. The increase in heap space is a temporary fix that we
may run into again as the repo grows - the out of memory error that
it prevents is an ongoing issue with the library itself
(https://github.com/egoist/tsup/issues/920).