ksearch: remove nested button
This PR removes a button which was nested inside of a hyperlink.
The button was only there for the onClick trigger, which can
simply be moved to the hyperlink. Further, it is a violation
of HTML5 to have HTML inside of a hyperlink, so this was
a WCAG violation.
J=TECHOPS-13454