v1.8.3: Add debouncing to filter search and autocomplete (#511)
This adds debouncing to the autocomplete and filter search calls. Currently, this is set to 200ms. This means that we will not fire autocomplete or filter search requests until 200ms has elapsed since the last keystroke, in order to save on unnecessary queries while the user is still actively typing. This number was chosen based on the average typing speed being roughly 4-5 characters per second.